Refactor: [BADA-239] coin 파일 위치 변경 #148
