+++ src/pages/mypage/coin-history/api/apis.ts
@@ -1,7 +1,7 @@
 import { END_POINTS } from '@/shared/api/endpoints';
 import { axiosInstance } from '@/shared/lib/axios/axiosInstance';
 
-import type { UserCoin } from '@/entities/coin/lib/types';
+import type { UserCoin } from '@/pages/mypage/coin-history/lib/types';
 
 export const getUserCoin = async (): Promise<UserCoin> => {
   const data: UserCoin = await axiosInstance.get(END_POINTS.MYPAGE.COIN);
