+++ src/pages/mypage/coin-history/lib/types.ts
