+++ src/pages/mypage/ui/DataUsageCardSection.tsx
@@ -8,8 +8,8 @@ import Link from 'next/link';
 import { ArcElement, Chart, Tooltip } from 'chart.js';
 import { Doughnut } from 'react-chartjs-2';
 
-import { useUserCoinQuery } from '@/entities/coin/model/queries';
 import { useUserDataUsageQuery } from '@/entities/data/model/queries';
+import { useUserCoinQuery } from '@/pages/mypage/coin-history/model/queries';
 import { ICONS } from '@/shared/config/iconPath';
 
 Chart.register(ArcElement, Tooltip);
